requirements/base.txt
Django==1.7.1
dj-database-url==0.3.0
django-dotenv==1.2
django-compressor==1.4
dj-static==0.0.6
